backend/pyproject.toml
[tool.poetry]
name = "hazard-reporting-system-be"
version = "0.1.0"
description = "A Hazard Reporting System"
authors = ["tasfiqulghani"]
license = "NA"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
django-redis = "^5.4.0"
celery = "^5.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-leaflet = "^0.30.1"
djangorestframework-api-key = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
